+++ src/controllers/FilesController.php
@@ -16,7 +16,10 @@ class FilesController extends Controller
 {
 
     /**
-     * @throws BadRequestHttpException
+     * @return Response
+     * @throws \yii\base\InvalidConfigException
+     * @throws \yii\web\BadRequestHttpException
+     * @throws \yii\web\MethodNotAllowedHttpException
      */
     public function actionImportFile(): Response
     {
@@ -35,21 +38,19 @@ class FilesController extends Controller
 
         // Has the file already been imported?
         $asset = EscapeDam::getInstance()->files->getImportedAsset($fileId, $fieldId, $elementId);
-
-        if (!$asset instanceof \craft\elements\Asset) {
+        if (empty($asset)) {
             try {
                 $asset = EscapeDam::getInstance()->files->importFile($fileId, $fieldId, $elementId, $siteId, $folderId);
             } catch (\Throwable $e) {
-                Craft::error('An error occurred when importing a DAM file: ' . $e->getMessage(), __METHOD__);
-                Craft::$app->getErrorHandler()->logException($e);
-                return $this->asErrorJson($e->getMessage());
+                Craft::error($e, __METHOD__);
+                return $this->asFailure($e->getMessage());
             }
         }
 
         return $this->asJson([
             'success' => true,
             'filename' => $asset->getFilename(),
-            'assetId' => (int)$asset->getId()
+            'assetId' => (int)$asset->getId(),
         ]);
         
     }