+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added the `escapedam/api/file-usage` endpoint
 ### Changed
 - For very large source images, a 2500px transform will be downloaded instead of the original when importing files from the DAM
+- Escape DAM now uses its own log file
+### Fixed
+- Fixed a bug where the selected "Import Location" wasn't selected in field setting forms
 
 ## 3.0.0 - 2024-08-05
 ### Added